Link for Website I worked on: Listeninghere.org
I worked on the database and back-end of development for this website during my time at University at Albany. I collaborated with other UAlbany students and a professor to make this possible. Used a multitude of programming sources.  It is still has some work that needs to be touched up.